requirements.txt
wheel
Pillow
gunicorn
django-dotenv
argon2-cffi
argon2-cffi-bindings
django
django-htmx
django-storages
boto3
psycopg2-binary
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pyjwt